+++ src/main/java/com/bttf/queosk/controller/KakaoLoginController.java
@@ -1,7 +1,7 @@
 package com.bttf.queosk.controller;
 
 import com.bttf.queosk.dto.KaKaoLoginForm;
-import com.bttf.queosk.dto.userdto.UserSignInDto;
+import com.bttf.queosk.dto.UserSignInDto;
 import com.bttf.queosk.service.KakaoLoginService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
