+++ src/main/java/com/bttf/queosk/entity/KakaoAuth.java
@@ -18,4 +18,15 @@ public class KakaoAuth {
     private String kakaoId;
     private String access;
     private String refresh;
+
+    public static KakaoAuth of(String email, String kakaoId,
+                            String refreshToken, String accessToken) {
+
+        return KakaoAuth.builder()
+                .email(email)
+                .kakaoId(kakaoId)
+                .refresh(refreshToken)
+                .access(accessToken)
+                .build();
+    }
 }
